Try to fix md5sum naming convention
Currently our standard build script creates a *.zip and a *.zip.md5sum.
TWRP looks for the md5 with the naming convention of *.zip.md5sum.

Jenkins builds create a *.zip and *.md5sum, which TWRP fails to locate
as a valid md5.  Have not tested on any other recoveries except TWRP,
on hammerhead

Change-Id: Ib7f3082f80bba2d2bcaa09051db9c357d1c0f0df